chore: move extension-* file to a subfolder extension
related to https://github.com/podman-desktop/podman-desktop/pull/10623
