+++ packages/main/src/development-mode-tracker.ts
@@ -16,9 +16,10 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
+import { ExtensionLoaderSettings } from '/@/plugin/extension/extension-loader-settings.js';
+
 import type { ConfigurationRegistry } from './plugin/configuration-registry.js';
 import { Emitter } from './plugin/events/emitter.js';
-import { ExtensionLoaderSettings } from './plugin/extension-loader-settings.js';
 
 // track the changes of the development mode and provides a single event to update
 export class DevelopmentModeTracker {
