+++ packages/main/src/plugin/extension/extension-loader-settings.ts
