+++ packages/main/src/plugin/extension/extension-loader.ts
@@ -35,58 +35,58 @@ import type { WebviewRegistry } from '/@/plugin/webview/webview-registry.js';
 import type { ExtensionError, ExtensionInfo, ExtensionUpdateInfo } from '/@api/extension-info.js';
 import type { ImageInspectInfo } from '/@api/image-inspect-info.js';
 
-import { securityRestrictionCurrentHandler } from '../security-restrictions-handler.js';
-import { getBase64Image, isLinux, isMac, isWindows } from '../util.js';
-import type { ApiSenderType } from './api.js';
-import type { PodInfo } from './api/pod-info.js';
-import type { AuthenticationImpl } from './authentication.js';
-import { CancellationTokenSource } from './cancellation-token.js';
-import type { Certificates } from './certificates.js';
-import type { CliToolRegistry } from './cli-tool-registry.js';
-import type { CommandRegistry } from './command-registry.js';
-import type { ConfigurationRegistry, IConfigurationNode } from './configuration-registry.js';
-import type { ContainerProviderRegistry } from './container-registry.js';
-import type { Context } from './context/context.js';
-import type { CustomPickRegistry } from './custompick/custompick-registry.js';
-import type { DialogRegistry } from './dialog-registry.js';
-import type { Directories } from './directories.js';
-import type { Event } from './events/emitter.js';
-import { Emitter } from './events/emitter.js';
-import { DEFAULT_TIMEOUT, ExtensionLoaderSettings } from './extension-loader-settings.js';
-import type { ExtensionWatcher } from './extension-watcher.js';
-import type { FilesystemMonitoring } from './filesystem-monitoring.js';
-import type { IconRegistry } from './icon-registry.js';
-import type { ImageCheckerImpl } from './image-checker.js';
-import type { ImageFilesRegistry } from './image-files-registry.js';
-import type { ImageRegistry } from './image-registry.js';
-import type { InputQuickPickRegistry } from './input-quickpick/input-quickpick-registry.js';
-import { InputBoxValidationSeverity, QuickPickItemKind } from './input-quickpick/input-quickpick-registry.js';
-import type { KubernetesClient } from './kubernetes/kubernetes-client.js';
-import type { MessageBox } from './message-box.js';
-import { ModuleLoader } from './module-loader.js';
-import type { OnboardingRegistry } from './onboarding-registry.js';
-import type { ProviderRegistry } from './provider-registry.js';
-import type { Proxy } from './proxy.js';
-import { createHttpPatchedModules } from './proxy-resolver.js';
-import { type SafeStorageRegistry } from './safe-storage/safe-storage-registry.js';
+import { securityRestrictionCurrentHandler } from '../../security-restrictions-handler.js';
+import { getBase64Image, isLinux, isMac, isWindows } from '../../util.js';
+import type { ApiSenderType } from '../api.js';
+import type { PodInfo } from '../api/pod-info.js';
+import type { AuthenticationImpl } from '../authentication.js';
+import { CancellationTokenSource } from '../cancellation-token.js';
+import type { Certificates } from '../certificates.js';
+import type { CliToolRegistry } from '../cli-tool-registry.js';
+import type { CommandRegistry } from '../command-registry.js';
+import type { ConfigurationRegistry, IConfigurationNode } from '../configuration-registry.js';
+import type { ContainerProviderRegistry } from '../container-registry.js';
+import type { Context } from '../context/context.js';
+import type { CustomPickRegistry } from '../custompick/custompick-registry.js';
+import type { DialogRegistry } from '../dialog-registry.js';
+import type { Directories } from '../directories.js';
+import type { Event } from '../events/emitter.js';
+import { Emitter } from '../events/emitter.js';
+import type { FilesystemMonitoring } from '../filesystem-monitoring.js';
+import type { IconRegistry } from '../icon-registry.js';
+import type { ImageCheckerImpl } from '../image-checker.js';
+import type { ImageFilesRegistry } from '../image-files-registry.js';
+import type { ImageRegistry } from '../image-registry.js';
+import type { InputQuickPickRegistry } from '../input-quickpick/input-quickpick-registry.js';
+import { InputBoxValidationSeverity, QuickPickItemKind } from '../input-quickpick/input-quickpick-registry.js';
+import type { KubernetesClient } from '../kubernetes/kubernetes-client.js';
+import type { MessageBox } from '../message-box.js';
+import { ModuleLoader } from '../module-loader.js';
+import type { OnboardingRegistry } from '../onboarding-registry.js';
+import type { ProviderRegistry } from '../provider-registry.js';
+import type { Proxy } from '../proxy.js';
+import { createHttpPatchedModules } from '../proxy-resolver.js';
+import { type SafeStorageRegistry } from '../safe-storage/safe-storage-registry.js';
 import {
   StatusBarAlignLeft,
   StatusBarAlignRight,
   StatusBarItemDefaultPriority,
   StatusBarItemImpl,
-} from './statusbar/statusbar-item.js';
-import type { StatusBarRegistry } from './statusbar/statusbar-registry.js';
-import type { NotificationRegistry } from './tasks/notification-registry.js';
-import type { ProgressImpl } from './tasks/progress-impl.js';
-import { ProgressLocation } from './tasks/progress-impl.js';
-import type { Telemetry } from './telemetry/telemetry.js';
-import type { TrayMenuRegistry } from './tray-menu-registry.js';
-import type { IDisposable } from './types/disposable.js';
-import { Disposable } from './types/disposable.js';
-import { TelemetryTrustedValue } from './types/telemetry.js';
-import { Uri } from './types/uri.js';
-import type { Exec } from './util/exec.js';
-import type { ViewRegistry } from './view-registry.js';
+} from '../statusbar/statusbar-item.js';
+import type { StatusBarRegistry } from '../statusbar/statusbar-registry.js';
+import type { NotificationRegistry } from '../tasks/notification-registry.js';
+import type { ProgressImpl } from '../tasks/progress-impl.js';
+import { ProgressLocation } from '../tasks/progress-impl.js';
+import type { Telemetry } from '../telemetry/telemetry.js';
+import type { TrayMenuRegistry } from '../tray-menu-registry.js';
+import type { IDisposable } from '../types/disposable.js';
+import { Disposable } from '../types/disposable.js';
+import { TelemetryTrustedValue } from '../types/telemetry.js';
+import { Uri } from '../types/uri.js';
+import type { Exec } from '../util/exec.js';
+import type { ViewRegistry } from '../view-registry.js';
+import { DEFAULT_TIMEOUT, ExtensionLoaderSettings } from './extension-loader-settings.js';
+import type { ExtensionWatcher } from './extension-watcher.js';
 
 /**
  * Handle the loading of an extension
