+++ packages/main/src/plugin/extension/extension-tsconfig-parser.spec.ts
