+++ packages/main/src/plugin/extensions-updater/extensions-updater.spec.ts
@@ -19,10 +19,10 @@
 import { app } from 'electron';
 import { afterEach, beforeEach, expect, test, vi } from 'vitest';
 
+import type { ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
 import type { ExtensionInfo } from '/@api/extension-info.js';
 
 import type { ConfigurationRegistry } from '../configuration-registry.js';
-import type { ExtensionLoader } from '../extension-loader.js';
 import type { ExtensionsCatalog } from '../extensions-catalog/extensions-catalog.js';
 import type { CatalogExtension } from '../extensions-catalog/extensions-catalog-api.js';
 import type { ExtensionInstaller } from '../install/extension-installer.js';
