+++ packages/main/src/plugin/icon-registry.spec.ts
@@ -20,8 +20,9 @@ import path from 'node:path';
 
 import { beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
+import type { AnalyzedExtension } from '/@/plugin/extension/extension-loader.js';
+
 import type { ApiSenderType } from './api.js';
-import type { AnalyzedExtension } from './extension-loader.js';
 import { IconRegistry } from './icon-registry.js';
 
 let iconRegistry: IconRegistry;
