+++ packages/main/src/plugin/install/extension-installer.ts
@@ -26,11 +26,11 @@ import { ipcMain } from 'electron';
 import * as tarFs from 'tar-fs';
 
 import type { Directories } from '/@/plugin/directories.js';
+import type { AnalyzedExtension, ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
 
 import type { ApiSenderType } from '../api.js';
 import type { ContributionManager } from '../contribution-manager.js';
 import { DockerDesktopContribution, DockerDesktopInstaller } from '../docker-extension/docker-desktop-installer.js';
-import type { AnalyzedExtension, ExtensionLoader } from '../extension-loader.js';
 import type { ExtensionsCatalog } from '../extensions-catalog/extensions-catalog.js';
 import type { ImageRegistry } from '../image-registry.js';
 import type { Telemetry } from '../telemetry/telemetry.js';
