+++ packages/main/src/plugin/onboarding-registry.spec.ts
@@ -20,11 +20,11 @@ import * as fs from 'node:fs';
 
 import { afterEach, beforeEach, describe, expect, expectTypeOf, test, vi } from 'vitest';
 
+import type { AnalyzedExtension } from '/@/plugin/extension/extension-loader.js';
 import type { OnboardingState } from '/@api/onboarding.js';
 
 import type { ApiSenderType } from './api.js';
 import { Context } from './context/context.js';
-import type { AnalyzedExtension } from './extension-loader.js';
 import { OnboardingRegistry } from './onboarding-registry.js';
 import type { Disposable } from './types/disposable.js';
